refactoring by making composable table component
