+++ src/App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { BrowserRouter, Route, Switch } from "react-router-dom";
-import TaskTable from './components/todaysTasks/TaskTable';
+import RecentTasks from './components/recent/RecentTasks';
 import AddTask from './components/forms/AddTask';
 import EditTask from './components/forms/EditTask';
 import Archive from './components/archive/Archive';
 import WelcomePage from "./components/welcomePage/WelcomePage";
 import Navbar from 'react-bootstrap/Navbar';
 import Nav from 'react-bootstrap/Nav';
-import Uncompleted from './components/todaysTasks/Uncompleted';
+import Uncompleted from './components/uncompleted/Uncompleted';
 import NotFound from './components/NotFound';
 
 export const AuthContext = React.createContext();
@@ -64,7 +64,7 @@ export default function App() {
             <WelcomePage isAuthed={authState.isAuthenticated} />
           )} />
           
-          {authState.isAuthenticated && <Route exact path="/tasks" component={TaskTable} />}
+          {authState.isAuthenticated && <Route exact path="/tasks" component={RecentTasks} />}
           {authState.isAuthenticated && <Route exact path="/Archive" component={Archive} />}
           {authState.isAuthenticated && <Route exact path="/tasks/add" component={AddTask} />}
           {authState.isAuthenticated && <Route path="/tasks/:id" component={EditTask} />}
