+++ src/index.css
@@ -59,9 +59,9 @@ body {
 }
 
 
-.align{
+.align {
   text-align: center;
-  margin-bottom: 2rem;
+  margin: 2rem;
 }
 
 .align2{
@@ -138,4 +138,8 @@ textarea {
   color: var(--white);
   background-color: var(--red);
   font-weight: bold;
+}
+
+.col-header {
+  color: var(--white);
 }
